fastapi/pyproject.toml
[tool.poetry]
name = "mind"
version = "0.1.0"
description = ""
authors = ["AlexBarciaJS <guillermoalexisbarcia@gmail.com>"]
readme = "README.md"
packages = [{include = "mind"}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = {extras = ["all"], version = "^0.115.11"}
uvicorn = ">=0.34.0,<0.35.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
sqlalchemy = ">=2.0.38,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
python-dotenv = "^1.0.1"
openai = "^1.65.2"
psycopg2 = "^2.9.10"
langchain = "^0.3.19"
langchain-openai = "^0.3.7"
requests = "^2.32.3"
langchain-community = "^0.3.18"
mysql-connector-python = "^9.2.0"
pydantic = "^2.10.6"
pypdf2 = "^3.0.1"
pymongo = "^4.11.2"
aiofiles = "^24.1.0"
motor = "^3.7.0"
pinecone = "^6.0.1"
fitz = "^0.0.1.dev2"
pymupdf = "^1.25.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
